class-33/demo/pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["yahyaalbashar <v.salvatore7.gs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.12.4"
pytz = "^2021.3"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
